backend/khuonmat/face_recognition_system/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.0.0
python-dateutil==2.8.2
mysql-connector-python==8.2.0 
torch==1.2.0
torchvision==0.4.0
easydict==1.9
tqdm==4.31.1
tensorboardX==2.0 